Find Parameter Range for Simulation

This function identifies the range of parameter values that produce simulation results within a specified relative range of a target objective value. It also returns the parameter value that minimizes the absolute error relative to the objective.

Usage

find_par_range(par, sim, obj, rel_rng)

Arguments

par

A numeric vector of parameter values.

sim

A numeric vector of simulation results corresponding to the parameter values.

obj

A numeric value representing the objective value to compare against.

rel_rng

A numeric value specifying the relative range for acceptable error.

Value

A list with the following components:

par_rng

A numeric vector of length 2 containing the minimum and maximum parameter values within the specified relative range of the objective.

par_val

A numeric value of the parameter that minimizes the absolute error relative to the objective.

x

A numeric vector of interpolated parameter values.

y

A numeric vector of interpolated simulation results corresponding to x.

Examples

if (FALSE) { # \dontrun{
find_par_range(sim[[1]], calculate_wyr(sim), 0.27, 0.05)
} # }