Plot Dotty

This function creates a dotty plot to visualize the relationship between parameter values and performance results.

Usage

plot_dotty(
  par,
  var,
  y_label = "y",
  n_col = 3,
  y_lim = NULL,
  y_inter = NULL,
  trend = FALSE,
  run_ids = NULL,
  low_up = FALSE
)

Arguments

par

A data frame of model parameter values for each simulation.

var

The model performance result vector or list of vectors to be plotted against parameter values.

y_label

(optional) Labels for the y-axis, either a single label or a vector corresponding to each variable if var is a list. Default y_label = 'y'.

n_col

(optional) Number of columns in the facet grid. Default n_col = 3.

y_lim

(optional) Limits for the y-axis. Default y_lim = NULL.

y_inter

(optional) Y-axis intercept value. Default y_inter = NULL.

trend

(optional) Logical, indicating whether to add a trend line. Default is trend = FALSE.

run_ids

(optional) A numeric vector of run IDs to be highlighted in the plot. Default is run_ids = NULL.

low_up

(optional) Logical, TRUE if whole possible parameter range should be used for x axis. Default is low_up = FALSE.

Value

A ggplot object representing the dotty plot.

Examples

if (FALSE) { # \dontrun{
# plot_dotty(my_data, 'performance', y_label = 'y_label')
} # }