Truncate a Vector with Ellipses

This function truncates a vector and adds ellipses ("...") to indicate omitted values. Truncation can be done from the left or both sides.

Usage

truncate(x, n, side = "left")

Arguments

x

A vector to be truncated.

n

An integer specifying the maximum number of elements to include before truncation.

side

A string indicating the side from which to truncate. Options are 'left' or 'both'. Default is 'left'.

Value

A string with the truncated vector. If truncation occurs, ellipses ("...") are added to indicate omitted values.